+++ src/main/java/com/cebi/service/AdminReportService.java
@@ -1,5 +1,6 @@
 package com.cebi.service;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -29,7 +30,7 @@ public interface AdminReportService {
 	byte[] downloadPdf(QueryData queryData, String bank, TellerMaster master) throws Exception;
 	public byte[] csvDownloadQueue(QueryData queryData, String bank, TellerMaster master) throws Exception;
 	public Map<String,Object> commonDownloadQueue(QueryData queryData, String bank,TellerMaster master) throws Exception;
-	public BigInteger getReportQueueStatusCount(String bank, String date);
+	public BigDecimal getReportQueueStatusCount(String bank, String date);
 	public  LinkedHashMap<String, Integer> getTotalCount(String bankCode);
 
 }
